taint/test/mochitest/test_fetch.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test HTML Taint Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let sink_names = [
        "fetch.url",
        "fetch.body"
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          is(report.detail.str, "http://mochi.test:8888/tests/taint/test/mochitest/hello", "Check sink string content");

          let flow = report.detail.str.taint[0].flow;
          is(flow[2].operation, sink_names[i]);

	  i += 1;
	  if (i >= sink_names.length) {
              SimpleTest.finish();
	  }
      }, false);

      function startTest() {
          let taint_string = String.tainted("http://mochi.test:8888/tests/taint/test/mochitest/hello",);

          fetch(taint_string);
          fetch("some_url", {
            method: 'POST',
            body: taint_string,
          });
      }

      
    </script>
  </head>
  <body onload="startTest()">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <p id="test"></p>
  </body>
</html>
